move logger under abstracts
