+++ test/coverlet.core.tests/Instrumentation/InstrumenterTests.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using Coverlet.Core.Helpers;
-using Coverlet.Core.Logging;
+using Coverlet.Core.Abstracts;
 using Coverlet.Core.Samples.Tests;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
